MAINT: Move src to _extensions
When splitting out _pywt, a number of new extension modules will be
created. This makes a sub-module for them to all go into.

This organization is by no means permanent, but will work for now. It's
only the private API anyway, so shouldn't be considered stable.
