+++ .gitignore
@@ -25,9 +25,9 @@ pip-log.txt
 
 # Project working files
 # Expanded Cython
-pywt/src/*.[ch]
+pywt/_extensions/*.[ch]
 # Expanded .src file
-pywt/src/c/wavelets.[ch]
+pywt/_extensions/c/wavelets.[ch]
 cythonize.dat
 pywt/__config__.py
 pywt/version.py
