+++ pywt/_extensions/c/common.c
