+++ pywt/_extensions/c/convolution.c
