+++ pywt/_extensions/c/convolution.template.c
