+++ pywt/_extensions/c/convolution.template.h
