+++ pywt/_extensions/c/wavelets_coeffs.h
