+++ pywt/_extensions/c/wavelets_coeffs.template.h
