+++ pywt/_extensions/common.pxd
