+++ pywt/_extensions/wavelet.pxd
