+++ pywt/_extensions/wavelets_list.pxi
