+++ pywt/_multilevel.py
@@ -12,8 +12,8 @@ from __future__ import division, print_function, absolute_import
 
 import numpy as np
 
-from ._pywt import Wavelet
-from ._pywt import dwt, idwt, dwt_max_level
+from ._extensions._pywt import Wavelet
+from ._extensions._pywt import dwt, idwt, dwt_max_level
 from ._multidim import dwt2, idwt2, dwtn, idwtn, _fix_coeffs
 
 __all__ = ['wavedec', 'waverec', 'wavedec2', 'waverec2', 'wavedecn',
